+++ src/components/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+  return (
+    <footer className="bg-blue-500 text-white text-sm py-4 text-center mt-auto">
+      &copy; 2023 full stack web development
+    </footer>
+  );
+};
+
+export default Footer;
